clean-up: rename + import fixes
