+++ mozzarellm/utils/__init__.py
@@ -1,1 +1,2 @@
 """Utility functions and data schemas for gene cluster analysis"""
+from .local_retrieval import local_knowledge_context_retriever
